Поддержка дробных количеств ингредиентов
- Изменен тип count с int на double в модели RecipeIngredient
- Обновлена схема API для поддержки чисел с плавающей точкой
- Исправлен контроллер для работы с дробными значениями

+++ lib/controller/recipe_ingredient_controller.dart
@@ -139,7 +139,7 @@ class RecipeIngredientController extends ResourceController {
         final recipeIngredient = RecipeIngredient()
           ..recipe = Recipe()..id = ingredient['recipeId'] as int
           ..ingredient = Ingredient()..id = ingredient['ingredientId'] as int
-          ..count = ingredient['count'] as int?;
+          ..count = ingredient['count'] as double?;
         
         final query = Query<RecipeIngredient>(transaction)
           ..values = recipeIngredient;
